requirements.txt
# Core FastAPI and database dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client for external APIs
httpx==0.25.2

# Background task scheduling
apscheduler==3.10.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1